+++ supervillain/generator/__init__.py
@@ -4,4 +4,4 @@ from .example import DoNothing
 from .metropolis import NeighborhoodUpdate
 import supervillain.generator.constraint
 import supervillain.generator.combining
-import supervillain.generator.worms
+import supervillain.generator.worm
